requirements.txt
streamlit>=1.22.0
pillow>=9.0.0
